% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{arrayT}
\alias{arrayT}
\title{Array T From Data Frame}
\usage{
arrayT(x, cols, replace_na = NA)
}
\arguments{
\item{x}{A data frame of tag releases. See details.}

\item{cols}{A list matching required names to column names in x. See details.}

\item{replace_na}{A number to replace NA values before they are removed.}
}
\value{
An array of class \code{arrayT}
}
\description{
Make array T from a data frame.
}
\details{
The data frame \code{x} must include columns for tag release area
and tag release time. Currently, both should be integer valued and start
indexing at one. If each row is corresponds to one tag release, no further
columns are needed. If each row summarizes a number of tag releases in one
area at one time, then \code{x} must include a column of release counts.
Optionally, an integer column indicating individual class (e.g. sex, length
class) can be included.

The list \code{cols} must include
\itemize{
\item \code{}
}
}
\examples{

}
