% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{barplot.mmmTMB}
\alias{barplot.mmmTMB}
\title{Barplot Movement Probabilities from an 'mmmTMB' Object}
\usage{
\method{barplot}{mmmTMB}(x = NULL, area_names = NULL)
}
\arguments{
\item{x}{An object of class \code{mmmTMB}}

\item{area_names}{\code{\link[=character]{character()}} Vector of area names}
}
\value{
An object with class \code{gg} and \code{ggplot}
}
\description{
Barplot Movement Probabilities from an 'mmmTMB' Object
}
\examples{

}
