% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_f}
\alias{check_f}
\title{Check Fishing Mortality Rate}
\usage{
check_f(f, null_ok = TRUE)
}
\arguments{
\item{f}{\code{\link[=array]{array()}} Fishing mortality rate}

\item{null_ok}{\code{\link[=logical]{logical()}} May f be \code{NULL}?}
}
\value{
\code{NULL}
}
\description{
Check Fishing Mortality Rate
}
