% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_h}
\alias{check_h}
\title{Check Instantaneous Tag Loss Rate}
\usage{
check_h(h)
}
\arguments{
\item{h}{\code{\link[=numeric]{numeric()}} Instantaneous tag loss rate}
}
\value{
\code{NULL}
}
\description{
Check Instantaneous Tag Loss Rate
}
