% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_k}
\alias{check_k}
\title{Negative Binomial Dispersioin Parameter}
\usage{
check_k(k, null_ok = TRUE)
}
\arguments{
\item{k}{\code{\link[=numeric]{numeric()}} Negative Binomial Dispersion. The variance is given
by var = mu + mu / k (NB1) or var = mu + mu^2 / k (NB2).}

\item{null_ok}{\code{\link[=logical]{logical()}} May k be \code{NULL}?}
}
\value{
\code{NULL}
}
\description{
Negative Binomial Dispersioin Parameter
}
