% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_l}
\alias{check_l}
\title{Check Tag Reporting Rate Data}
\usage{
check_l(l, null_ok = TRUE)
}
\arguments{
\item{null_ok}{\code{\link[=logical]{logical()}} May l be \code{NULL}?}

\item{z}{\code{\link[=array]{array()}} Tag reporting rate data}
}
\value{
\code{NULL}
}
\description{
Check Tag Reporting Rate Data
}
