% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_m}
\alias{check_m}
\title{Check Natural Mortality Rate}
\usage{
check_m(m, null_ok = TRUE)
}
\arguments{
\item{m}{\code{\link[=numeric]{numeric()}} Natural mortality rate}

\item{null_ok}{\code{\link[=logical]{logical()}} May m be \code{NULL}?}
}
\value{
\code{NULL}
}
\description{
Check Natural Mortality Rate
}
