% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_p}
\alias{check_p}
\title{Check Movement Parameters}
\usage{
check_p(p, null_ok = TRUE)
}
\arguments{
\item{p}{\code{\link[=array]{array()}} Movement parameters}

\item{null_ok}{\code{\link[=logical]{logical()}} May p be \code{NULL}?}
}
\value{
\code{NULL}
}
\description{
Check Movement Parameters
}
