% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_settings}
\alias{check_settings}
\title{Check Settings Argument for mmmFit()}
\usage{
check_settings(settings)
}
\arguments{
\item{settings}{\code{\link[=list]{list()}}}
}
\value{

}
\description{
Check Settings Argument for mmmFit()
}
\examples{

}
