% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_u}
\alias{check_u}
\title{Check Initial Tag Loss Rate (Proportion)}
\usage{
check_u(u)
}
\arguments{
\item{u}{\code{\link[=numeric]{numeric()}} Initial tag loss rate (proportion)}
}
\value{
\code{NULL}
}
\description{
Check Initial Tag Loss Rate (Proportion)
}
