% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_y}
\alias{check_y}
\title{Check Tag Recovery Data}
\usage{
check_y(y)
}
\arguments{
\item{y}{\link[=mmmTags]{matrix()} Tag recovery data}
}
\value{
\code{NULL}
}
\description{
Check Tag Recovery Data
}
