% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_z}
\alias{check_z}
\title{Check Index Matrix}
\usage{
check_z(z)
}
\arguments{
\item{z}{\link[=mmmIndex]{matrix()} Square binary index matrix}
}
\value{
\code{NULL}
}
\description{
Check Index Matrix
}
