% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_bias_results}
\alias{create_bias_results}
\title{Create Fishing Mortality Bias Results}
\usage{
create_bias_results(log_b, log_b_cov, estimate)
}
\arguments{
\item{log_b}{\code{\link[=numeric]{numeric()}} Vector.}

\item{log_b_cov}{\code{\link[=matrix]{matrix()}} Covariances.}

\item{estimate}{\code{\link[=logical]{logical()}} Estimate or return NULL.}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Create Fishing Mortality Bias Results
}
