% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_dispersion_results}
\alias{create_dispersion_results}
\title{Create Dispersion Results}
\usage{
create_dispersion_results(log_k, log_k_se, estimate)
}
\arguments{
\item{log_k}{\code{\link[=numeric]{numeric()}} Scalar estimate.}

\item{log_k_se}{\code{\link[=numeric]{numeric()}} Scalar estimate.}

\item{estimate}{\code{\link[=logical]{logical()}} Estimate negative binomial dispersion?}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Create Dispersion Results
}
