% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_fishing_results}
\alias{create_fishing_results}
\title{Create Fishing Rate Results}
\usage{
create_fishing_results(
  vlogit_exp_neg_tf,
  mlogit_exp_neg_tf_cov,
  results_step,
  nft,
  nfa,
  nfg,
  estimate
)
}
\arguments{
\item{vlogit_exp_neg_tf}{\code{\link[=numeric]{numeric()}} Vector estimates.}

\item{mlogit_exp_neg_tf_cov}{\code{\link[=matrix]{matrix()}} Covariances.}

\item{results_step}{\code{\link[=integer]{integer()}} Results step as multiple of tag step.}

\item{nft}{\code{\link[=integer]{integer()}} Number of fishing rate time steps.}

\item{nfa}{\code{\link[=integer]{integer()}} Number of fishing rate areas.}

\item{estimate}{\code{\link[=logical]{logical()}} Estimate fishing mortality rate(s)?}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Create Fishing Rate Results
}
