% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_index_vector}
\alias{create_index_vector}
\title{Create Index Vector}
\usage{
create_index_vector(a, b, from = 0)
}
\arguments{
\item{a}{\code{\link[=integer]{integer()}} Length of object to index}

\item{b}{\code{\link[=integer]{integer()}} Length of index vector}

\item{from}{\code{\link[=integer]{integer()}} Value from which to index}
}
\value{
An \code{\link[=integer]{integer()}} vector
}
\description{
Create Index Vector
}
\examples{
# a == 1
create_index_vector(1, 3) # Repeated

# a == b
create_index_vector(3, 3) # Sequence

# a < b
create_index_vector(2, 3) # Truncated sequence of repeated values
create_index_vector(2, 4) # Sequence of repeated values
create_index_vector(5, 9) # Truncated sequence of repeated values

# a > b
create_index_vector(5, 3) # Truncated sequence

}
