% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_mortality_results}
\alias{create_mortality_results}
\title{Create Natural Mortality Results}
\usage{
create_mortality_results(
  logit_exp_neg_m,
  logit_exp_neg_m_se,
  results_step,
  estimate
)
}
\arguments{
\item{logit_exp_neg_m}{\code{\link[=numeric]{numeric()}} Scalar estimate.}

\item{logit_exp_neg_m_se}{\code{\link[=numeric]{numeric()}} Scalar estimate.}

\item{results_step}{\code{\link[=integer]{integer()}} Results step as multiple of tag step.}

\item{estimate}{\code{\link[=logical]{logical()}} Estimate natural mortality?}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Create Natural Mortality Results
}
