% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_movement_parameters.default}
\alias{create_movement_parameters.default}
\title{Default Create Movement Parameter Array}
\usage{
\method{create_movement_parameters}{default}(x, ...)
}
\arguments{
\item{x}{A vector of parameter values or array of movement rates}

\item{...}{Other arguments for class specific methods}
}
\value{

}
\description{
Default Create Movement Parameter Array
}
\examples{

}
