% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_movement_parameters.numeric}
\alias{create_movement_parameters.numeric}
\title{Create Movement Parameter Array from a Vector of Parameters}
\usage{
\method{create_movement_parameters}{numeric}(x, d = NULL, t = FALSE)
}
\arguments{
\item{x}{Numeric vector. Movement parameters.}

\item{d}{Integer vector. Dimensions based on source of \code{x}. From
\code{R} use \code{c(npt, np, ng)} or from \code{TMB} use
\code{c(np, npt, ng)}.}

\item{t}{Logical. Transpose the first two dimensions of the parameter
array?}
}
\value{
An array.
}
\description{
Create Movement Parameter Array from a Vector of Parameters
}
\examples{

# aP
x <- rnorm(24, 0, 1)
d <- c(6, 4, 2)
aP <- create_movement_parameters(x, d)

# taP
x <- rnorm(24, 0, 1)
d <- c(6, 4, 2)
taP <- create_movement_parameters(x, d, TRUE)

}
