% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_movement_parameters}
\alias{create_movement_parameters}
\title{Generic Create Movement Parameter Array}
\usage{
create_movement_parameters(x, d = NULL, t = FALSE)
}
\arguments{
\item{x}{A vector of parameter values or array of movement rates}

\item{d}{A vector of array dimensions (only used for vector method)}

\item{t}{Logical. Transpose the first two dimensions of the parameter
array?}
}
\value{
An array
}
\description{
Generic Create Movement Parameter Array
}
\examples{

}
