% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_movement_rates}
\alias{create_movement_rates}
\title{Create Movement Rate Array}
\usage{
create_movement_rates(a, m)
}
\arguments{
\item{a}{Numeric array. Movement parameters. (Not transposed).}

\item{m}{Integer matrix. Movement indexes. (Not transposed).}
}
\value{
\code{\link[=array]{array()}} Movement rates
}
\description{
Create Movement Rate Array
}
\examples{

}
