% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{create_release_matrix}
\alias{create_release_matrix}
\title{Create Tag Release Matrix}
\usage{
create_release_matrix(
  n = 1000,
  steps = 30,
  areas = 3L,
  groups = 1L,
  errors = FALSE
)
}
\arguments{
\item{n}{\code{\link[=integer]{integer()}} Mean tag release count per stratum}

\item{steps}{\code{\link[=integer]{integer()}} Number of time steps}

\item{areas}{\code{\link[=integer]{integer()}} Number of areas}

\item{groups}{\code{\link[=integer]{integer()}} Number of groups}

\item{errors}{\code{\link[=logical]{logical()}} Poisson distributed tag release counts?}
}
\value{
An integer matrix
}
\description{
Create Tag Release Matrix
}
\examples{

m1 <- create_release_matrix()
m2 <- create_release_matrix(groups = 3)
m3 <- create_release_matrix(errors = TRUE)

}
