% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{get_convergence_diagnostics}
\alias{get_convergence_diagnostics}
\title{Get Convergence Diagnostics from TMB Model Fit}
\usage{
get_convergence_diagnostics(sd_report)
}
\arguments{
\item{sd_report}{\code{\link[=sdreport]{sdreport()}} See \code{\link[TMB:sdreport]{TMB::sdreport()}}.}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Get convergence diagnostics from a TMB model fit.
This function is from the sdmTMB' package at
github.com/pbs-assess/sdmTMB/blob/master/R/fit.R
}
\examples{

}
