% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{invlogit}
\alias{invlogit}
\title{Inverse Logit}
\usage{
invlogit(y)
}
\arguments{
\item{y}{\code{\link[=numeric]{numeric()}} Finite or infinite scalar.}
}
\value{
The inverse of the log odds ratio
}
\description{
Inverse Logit
}
\examples{
invlogit(Inf)
invlogit(0)
invlogit(-Inf)
invlogit(matrix(c(-Inf, 0, 0, Inf), nrow = 2))

}
