% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lineplot}
\alias{lineplot}
\title{Lineplot Movement Rates from an 'mmmTMB' Object}
\usage{
lineplot(
  x = NULL,
  x_name = "Year",
  y_name = "Annual Movement Rate",
  x_breaks = NULL,
  x_labels = NULL,
  area_names = NULL,
  class_names = NULL,
  error_alpha = 1,
  error_width = 0.2,
  legend_title = "Length Class",
  legend_position = "right"
)
}
\arguments{
\item{x}{An object of class \code{mmmTMB}}

\item{x_name}{\code{\link[=character]{character()}} Axis name}

\item{y_name}{\code{\link[=character]{character()}} Axis name}

\item{x_breaks}{\code{\link[=numeric]{numeric()}} Axis breaks}

\item{x_labels}{\code{\link[=character]{character()}} Axis labels}

\item{area_names}{\code{\link[=character]{character()}} Vector of area names}

\item{class_names}{\code{\link[=character]{character()}} Vector of class names}

\item{error_alpha}{\code{\link[=numeric]{numeric()}} Error bar alpha}

\item{error_width}{\code{\link[=numeric]{numeric()}} Error bar width}

\item{legend_title}{\code{\link[=character]{character()}} Legend title}

\item{legend_position}{\code{\link[=character]{character()}} Legend position}
}
\value{
An object with class \code{gg} and \code{ggplot}
}
\description{
Lineplot Movement Rates from an 'mmmTMB' Object
}
\examples{

}
