% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{logit}
\alias{logit}
\title{Logit (Log Odds Ratio)}
\usage{
logit(x)
}
\arguments{
\item{x}{Numeric between zero and one(inclusive)}
}
\value{
The log of the odds ratio
}
\description{
Logit (Log Odds Ratio)
}
\examples{

}
