% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{matrix_power}
\alias{matrix_power}
\title{Matrix Power}
\usage{
matrix_power(x, n)
}
\arguments{
\item{x}{\code{\link[=matrix]{matrix()}} Numeric matrix}

\item{n}{\code{\link[=integer]{integer()}} Matrix power}
}
\value{
\code{\link[=matrix]{matrix()}}
}
\description{
Matrix Power
}
\examples{

x <- matrix(c(1,2,3,4), nrow = 2)
n <- 2
matrix_power(x, n)

}
