% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmmControl}
\alias{mmmControl}
\title{Optimization control options}
\usage{
mmmControl(eval.max = 10000, iter.max = 10000, ...)
}
\arguments{
\item{eval.max}{Maximum number of evaluations of the objective
function allowed.}

\item{iter.max}{Maximum number of iterations allowed.}

\item{...}{Anything else. See the 'Control parameters' section of
\code{stats::nlminb()}.}
}
\description{
Any arguments to pass to \code{stats::nlminb()}.
}
