% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmmFit}
\alias{mmmFit}
\title{Fit a Markov Movement Model}
\usage{
mmmFit(
  data,
  parameters = NULL,
  map = NULL,
  settings = mmmSet(),
  control = mmmControl()
)
}
\arguments{
\item{data}{\code{\link[=list]{list()}} Tag and other data. See Details.}

\item{parameters}{\code{\link[=list]{list()}} Initial parameter values. See Details.}

\item{map}{\code{\link[=list]{list()}} Values to override \code{map} defaults.
See \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}.}

\item{settings}{\code{\link[=list]{list()}} Values that help define the model.
See \code{\link[=mmmSet]{mmmSet()}}.}

\item{control}{\code{\link[=list]{list()}} Optimization control options to pass to
\code{\link[stats:nlminb]{stats::nlminb()}}. See \code{\link[=mmmControl]{mmmControl()}}.}
}
\value{
A list of class \code{mmmFit}.
}
\description{
Estimate movement rates from single-mark single-recapture
tagging data.
}
\details{
The \code{\link[=list]{list()}} argument \code{data} must contain:
\itemize{
\item \code{x} An integer matrix of tag release counts formatted
as described in \code{\link[=mmmTags]{mmmTags()}}.
\item \code{y} An integer matrix of tag recovery counts formatted
as described in \code{\link[=mmmTags]{mmmTags()}}.
\item \code{z} A square binary index matrix that indicates allowed
direct movement between areas (from rows to columns) from one
time step to the next. Ones off the diagonal indicate allowed
direct movement. Zeros off the diagonal represent disallowed
direct movement. Numbers on the diagonal are ignored because
self-movement is always allowed. See \code{\link[=mmmIndex]{mmmIndex()}}.
\item \code{h} The (positive) scalar instantaneous tag loss rate.
\item \code{u} The scalar proportion of tags lost during release.
}
The list argument \code{data} may optionally contain:
\itemize{
\item \code{l} A matrix of tag reporting rates (proportions).
Defaults to one if omitted (full reporting). See \code{\link[=mmmRates]{mmmRates()}}.
\item \code{w} A matrix of fishing mortality rate weights. Useful
when the tag time step is finer than the fishing rate step.
Defaults to one (equal fishing rates across tag time steps
within a fishing rate time step). See \code{\link[=mmmWeights]{mmmWeights()}}.
\item \code{f} A matrix of fishing mortality rates. Estimated as
parameter(s) if omitted. See \code{parameters} and \code{settings}.
Also see \code{\link[=mmmRates]{mmmRates()}}.
\item \code{m} The scalar instantaneous natural mortality rate.
Estimated if omitted. One or both of \code{f} and \code{m} must be
provided. See \code{parameters}.
}
The list argument \code{parameters} may optionally contain the initial
values:
\itemize{
\item \code{p} An array of movement parameters. See TBD.
\item \code{f} A matrix of fishing mortality rates. See \code{settings}.
\item \code{m} The scalar instantaneous natural mortality rate.
\item \code{b} The vector fishing mortality bias by group.
\item \code{k} The scalar negative binomial dispersion.
}
}
\examples{

}
\author{
Luke A. Rogers
}
