% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{mmmGroup}
\alias{mmmGroup}
\title{Map Values to Groups}
\usage{
mmmGroup(x, groups)
}
\arguments{
\item{x}{\code{\link[=atomic]{atomic()}} Vector of values.}

\item{groups}{\code{\link[=list]{list()}} Named vectors of group elements.}
}
\value{
\code{\link[=integer]{integer()}} Vector.
}
\description{
Map Values to Groups
}
\examples{

# Numeric
x <- c(1, 4, 2, 3, NA, 2, 7)
y <- list(a = 1:2, b = 3:4)
g1 <- mmmGroup(x, y)
# Character
x <- c("M", "F", "F", NA, "M", "N", "F")
y <- list(m = "M", f = "F")
g2 <- mmmGroup(x, y)
# Factor
x <- factor(c("L", "S", "M", NA, "M", "H"))
y <- list(s = "S", m = "M", l = "L")
g3 <- mmmGroup(x, y)

}
