% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{mmmRates}
\alias{mmmRates}
\title{Prepare Rate Data}
\usage{
mmmRates(
  x,
  cols = NULL,
  year_lims = NULL,
  area_names = NULL,
  replicates = 1,
  instantaneous = FALSE
)
}
\arguments{
\item{x}{A data frame of named columns. See details.}

\item{cols}{A list of named character elements. See details.}
}
\value{

}
\description{
Prepare Rate Data
}
\examples{
year <- rep(2010:2016, 3)
area <- rep(c(1,2,3), each = 7)
rate <- rep(seq(0.1, 0.5, length.out = 7), 3)
x <- data.frame(year = year, area = area, rate = rate)
r1 <- mmmRates(x)

}
