% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmmSet}
\alias{mmmSet}
\title{Settings for \code{mmmFit()}}
\usage{
mmmSet(
  error_family = c(nb1 = 1, poisson = 0),
  min_liberty = 1,
  max_liberty = 1000,
  time_varying = 0,
  time_process = c(none = 0, rw = 1),
  cycle_length = 0,
  block_length = 0,
  results_step = 1,
  nlminb_loops = 5,
  newton_iters = 5,
  openmp_cores = 1
)
}
\arguments{
\item{error_family}{Integer. \code{0} = Poisson; \code{1} = NB1;
\code{2} = NB2.}

\item{max_liberty}{Integer. Max time steps at liberty before
recapture.}

\item{time_varying}{Integer. \code{0} = None; \code{1} = Time varying
movement rates.}

\item{time_process}{Integer. \code{0} = None; \code{1} = Negative
binomial (NB1).}

\item{cycle_length}{Integer. Cycle length or \code{0} for no cycle.}

\item{block_length}{Integer. Block length for movement rate.}

\item{results_step}{Integer. Results time step as a multiple of the data
time step.}

\item{nlminb_loops}{Integer. Number of times to run \code{\link[stats:nlminb]{stats::nlminb()}}
optimization.}

\item{openmp_cores}{Integer. Number of parallel cores for TMB.}

\item{newton_steps}{Integer. Number of Newton optimization steps.}
}
\value{
A list of settings for \code{mmmFit()}
}
\description{
Settings for \code{mmmFit()}
}
\examples{

}
