% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{mmmTags}
\alias{mmmTags}
\title{Prepare Tag Data}
\usage{
mmmTags(x, y, cols = NULL, time_step = NULL, date_lims = NULL)
}
\arguments{
\item{x}{A data frame of named columns. See details.}

\item{y}{A data frame of named columns. See details.}

\item{cols}{A list of named character elements. See details.}

\item{time_step}{A character string. One of \code{year}, \code{month}, or
\code{day}.}

\item{date_lims}{A character vector with two elements. "\%Y-\%M-\%D".}
}
\value{

}
\description{
Prepare Tag Data
}
\examples{

}
