% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmmTMB}
\alias{mmmTMB}
\title{Fit a Markov Movement Model to Mark-Recapture Data}
\usage{
mmmTMB(data_list, ...)
}
\arguments{
\item{data_list}{\code{\link[=list]{list()}} A named list of data objects (see details)}

\item{structure_list}{\code{\link[=list]{list()}} A named list specifying model structure (see
details)}

\item{optimizer_list}{\code{\link[=list]{list()}} A named list of optimizer arguments (see
details)}

\item{parameter_list}{\code{\link[=list]{list()}} An optional named list of initial values
(see details)}
}
\value{
An object of class \code{mmmTMB}
}
\description{
Estimate movement probabilities among areas by fitting a Markov
movement model to mark release and recovery data.

Array dimensions are abbreviated for space and are given by:
\itemize{
\item{\code{nt}: Count of time steps in the data}
\item{\code{na}: Count of areas in the data}
\item{\code{ng}: Count of group categories}
}
}
\details{
The argument \code{data_list} must be a named \code{list()} of objects:
\itemize{
\item{\code{released_3d} an \code{array()} with \code{dim = c(nt, na, ng)}}
\item{\code{recovered_5d} an \code{array()} with
\code{dim = c(nt, nt, na, na, ng)}}
\item{\code{capture_rate_2d} an \code{array()} with \code{dim = c(na, nt)}}
\item{\code{report_ratio_2d} an \code{array()} with \code{dim = c(na, nt)}}
\item{\code{template_2d} an \code{array()} with \code{dim = c(na, na)}}
\item{\code{tag_loss_rate} a \code{numeric()} with \code{length = 1}}
\item{\code{imm_loss_ratio} a \code{numeric()} with \code{length = 1}}
}

The argument \code{structure_list} must be a named \code{list()} of objects:
\itemize{
\item{\code{recapture_delay} \code{integer()} Minimum delay between
release and recovery}
\item{\code{error_family} \code{integer()} One of 0: Poisson; 1: NB1;
or 2: NB2}
\item{\code{result_units} \code{integer()} Results time step as multiple
of data time step}
}

The argument \code{optimizer_list} must be a named \code{list()} of objects:
\itemize{
\item{\code{newton_steps} \code{integer()} Number of Newton optimization
steps}
\item{\code{nlminb_loops} \code{integer()} Number of times to run
\code{\link[stats:nlminb]{stats::nlminb()}} optimization.}
\item{\code{nlminb_control} \code{list()} see \code{mmmTMBcontrol()}}
\item{\code{openmp_cores} \code{integer()} Number of cores for TMB}
}

The argument \code{parameter_list} must be \code{NULL} or a named
\code{list()} of objects:
\itemize{
\item{TBD...}
}
}
\examples{

}
