% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmmTMBcontrol}
\alias{mmmTMBcontrol}
\title{Optimization control options}
\usage{
mmmTMBcontrol(eval.max = 10000, iter.max = 10000, ...)
}
\arguments{
\item{eval.max}{\link{numeric(1)} Maximum number of evaluations of the objective
function allowed.}

\item{iter.max}{\link{numeric(1)} Maximum number of iterations allowed.}

\item{...}{Anything else. See the 'Control parameters' section of
\code{\link[stats:nlminb]{stats::nlminb()}}.}
}
\description{
Any arguments to pass to \code{\link[stats:nlminb]{stats::nlminb()}}.
}
