% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{mmmWeights}
\alias{mmmWeights}
\title{Create Monthly Weighting for Annual Fishing Mortality}
\usage{
mmmWeights(tags, rate_step = "year")
}
\arguments{
\item{tags}{An \code{mmmTags} object. See \code{mmmTags()}.}

\item{rate_step}{String. Currenlty implemented for \code{"month"} only.}
}
\value{
A matrix of monthly weights for fishing mortality rates.
}
\description{
Create Monthly Weighting for Annual Fishing Mortality
}
\examples{

}
