% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mmmTMB}
\alias{plot.mmmTMB}
\title{Plot Movement Probabilities from an 'mmmTMB' Object}
\usage{
\method{plot}{mmmTMB}(
  x = NULL,
  class = 1,
  xlab = "Area To",
  ylab = "Area From",
  area_names = NULL,
  font_size_probs = 3,
  font_nudge_probs = 1,
  font_size_stderr = 1.5,
  font_nudge_stderr = 1,
  legend_name = "Estimate",
  legend_position = "right"
)
}
\arguments{
\item{x}{An object of class \code{mmmTMB}}

\item{class}{\code{\link[=integer]{integer()}} Release class}

\item{xlab}{\code{\link[=character]{character()}} The x label of the plot}

\item{ylab}{\code{\link[=character]{character()}} The y label of the plot}

\item{area_names}{\code{\link[=character]{character()}} Vector of area names}

\item{font_size_probs}{\code{\link[=numeric]{numeric()}} Font size for probabilities}

\item{font_nudge_probs}{\code{\link[=numeric]{numeric()}} Upward displacement for probabilities}

\item{font_size_stderr}{\code{\link[=numeric]{numeric()}} Font size for standard errors}

\item{font_nudge_stderr}{\code{\link[=numeric]{numeric()}} Downward displacement for standard errors}

\item{legend_name}{\code{\link[=character]{character()}} Legend title}

\item{legend_position}{\code{\link[=character]{character()}} Legend position}
}
\value{
An object with class \code{gg} and \code{ggplot}
}
\description{
Plot Movement Probabilities from an 'mmmTMB' Object
}
\examples{

}
