% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sensibly_round}
\alias{sensibly_round}
\title{Sensibly Round}
\usage{
sensibly_round(x, digits)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}} Value to round}

\item{digits}{\link{numeric} Number of digits}
}
\value{
\code{\link[=character]{character()}}
}
\description{
Sensibly Round
}
\examples{
sensibly_round(c(0.0005, 0.0010, 0.0015), 3)

}
