% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{subset_by_name}
\alias{subset_by_name}
\title{Subset Numeric Vector or Matrix by Rowname, Colname, or Name}
\usage{
subset_by_name(x, x_names)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}} or \code{\link[=matrix]{matrix()}}}

\item{x_names}{\code{\link[=character]{character()}}}
}
\value{

}
\description{
Subset Numeric Vector or Matrix by Rowname, Colname, or Name
}
\examples{

}
