% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_data}
\alias{check_data}
\title{Check Data Argument for mmmFit()}
\usage{
check_data(data = NULL)
}
\arguments{
\item{data}{\code{\link[=list]{list()}}}
}
\value{

}
\description{
Check Data Argument for mmmFit()
}
\examples{

}
