% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_parameters}
\alias{check_parameters}
\title{Check Parameters Argument for mmmFit()}
\usage{
check_parameters(parameters)
}
\arguments{
\item{parameters}{\code{\link[=list]{list()}}}
}
\value{

}
\description{
Check Parameters Argument for mmmFit()
}
\examples{

}
