% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_movement_parameters.array}
\alias{create_movement_parameters.array}
\title{Create Movement Parameter Array from a Movement Rate Array}
\usage{
\method{create_movement_parameters}{array}(x, d = NULL, t = FALSE)
}
\arguments{
\item{x}{Numeric Array. aK.}

\item{d}{Placeholder.}

\item{t}{Logical. Transpose the first two dimensions of the parameter
array?}
}
\value{
An array
}
\description{
Create Movement Parameter Array from a Movement Rate Array
}
\examples{

}
