% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_movement_parameters}
\alias{create_movement_parameters}
\title{Generic Create Movement Parameter Array}
\usage{
create_movement_parameters(x, ...)
}
\arguments{
\item{x}{A vector of parameter values or array of movement rates}
}
\value{
An array
}
\description{
Generic Create Movement Parameter Array
}
\examples{

}
