% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_movement_parameters}
\alias{create_movement_parameters}
\title{Create Movement Parameter Array}
\usage{
create_movement_parameters(r, z = NULL)
}
\arguments{
\item{r}{Numeric Array. aK.}

\item{z}{Integer matrix. Movement indexes. dim = c(pa, ca)}
}
\value{
\code{\link[=array]{array()}}
}
\description{
Create Movement Parameter Array
}
\examples{

# 2x2
x <- array(c(0.9, 0.2, 0.1, 0.8, 0.7, 0.4, 0.3, 0.6), dim = c(2, 2, 2, 1))
z <- array(c(0, 1, 1, 0), dim = c(2, 2))
p <- create_movement_parameters(x, m)
r1 <- create_movement_rates(a, m)
all(x - r < 1e-12)

# 3x3 with missing adjacent rates
v <- c(0.9,0.1,0.3,0,0.8,0,0.1,0.1,0.7,0.6,0.2,0.6,0,0.5,0,0.4,0.3,0.4)
x <- array(v, dim = c(3,3,2,1))
z <- array(c(0,1,1,0,0,0,1,1,0), dim = c(3,3))
p <- create_movement_parameters(x, m)
r <- create_movement_rates(a, m)
all(x - r < 1e-12)

}
