% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_movement_rates}
\alias{create_movement_rates}
\title{Create Movement Rate Array}
\usage{
create_movement_rates(a, m, t = FALSE)
}
\arguments{
\item{a}{Numeric array. Movement parameters.}

\item{m}{Integer matrix. Movement indexes.}

\item{t}{Are \code{a} and \code{m} transposes?}
}
\value{

}
\description{
Create Movement Rate Array
}
\examples{

}
