% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_movement_results}
\alias{create_movement_results}
\title{Create Movement Results}
\usage{
create_movement_results(v, m, np, npt, ng, mi, pow, draws)
}
\arguments{
\item{v}{\code{\link[=numeric]{numeric()}} Vector of movement parameters in transpose order.}

\item{m}{\code{\link[=matrix]{matrix()}} Matrix of movement parameter covariances in
transpose order.}

\item{np}{\code{\link[=integer]{integer()}} Number of movement parameters per step and group.}

\item{npt}{\code{\link[=integer]{integer()}} Number of movement parameter time steps.}

\item{ng}{\code{\link[=integer]{integer()}} Number of groups.}

\item{mi}{\code{\link[=matrix]{matrix()}} Matrix index. See \code{\link[=mmmIndex]{mmmIndex()}}.}

\item{pow}{\code{\link[=integer]{integer()}} Results step as a multiple of tag time step.
A matrix power.}

\item{draws}{\code{\link[=integer]{integer()}} Number of draws from which to bootstrap
movement result standard errors.}
}
\value{
A \code{\link[=list]{list()}}
}
\description{
Create Movement Results
}
\examples{

}
