% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{logit}
\alias{logit}
\title{Logit (Log Odds Ratio)}
\usage{
logit(x)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}} Value between zero and one (inclusive).}
}
\value{
\code{\link[=numeric]{numeric()}} The log of the odds ratio.
}
\description{
Logit (Log Odds Ratio)
}
\examples{
logit(0)
logit(0.5)
logit(1)
logit(matrix(c(0, 0.5, 0.5, 1), nrow = 2))

}
