% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{matrix_power}
\alias{matrix_power}
\title{Matrix Power}
\usage{
matrix_power(x, pow)
}
\arguments{
\item{x}{\code{\link[=matrix]{matrix()}} Square numeric matrix.}

\item{pow}{\code{\link[=integer]{integer()}} Matrix power.}
}
\value{
\code{\link[=matrix]{matrix()}}
}
\description{
Matrix Power
}
\examples{

x <- matrix(c(1,2,3,4), nrow = 2)
pow <- 2
matrix_power(x, pow)

}
