% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmmFit}
\alias{mmmFit}
\title{Fit a Markovian Movement Model}
\usage{
mmmFit(
  data,
  parameters = NULL,
  random = NULL,
  map = NULL,
  settings = mmmSet(),
  control = mmmControl()
)
}
\arguments{
\item{data}{A list of named data objects. See Details.}

\item{parameters}{A list of initial parameter values. See Details.}

\item{random}{A character vector of parameters to estimate as random effects.}

\item{map}{A list of \code{map} values to override defaults.
See \code{TMB::MakeADFun()}.}

\item{settings}{A list of values that define the model.
See \code{mmmSet()}.}

\item{control}{A list of optimization control options to pass to
\code{stats::nlminb()}. See \code{mmmControl()}.}
}
\value{
A list of class \code{mmmFit}.
}
\description{
Estimate movement rates
}
\details{
The list argument \code{data} must contain
\itemize{
\item{\code{tags} a list of class \code{mmmTags} (recommended: see
\code{mmmTags()}) OR both \code{mT} and \code{mR} integer matrices
of tag release and recovery counts, respectively, formatted as
described in \code{mmmTags()}}
\item{\code{mI} a square binary index matrix representing movement
between areas (from rows to columns). Ones represent movement that
is allowed from one time step to the next. Zeros off the diagonal
represent disallowed movement. Numbers on the diagonal are ignored
because self-movement is always allowed.}
}
The list argument \code{data} may optionally contain
\itemize{
\item{\code{mL} a matrix of tag reporting rates (proportions)
in which the time step is given by the row and the area is
given by the column. Defaults to ones (full reporting).}
\item{mW}
}
}
\examples{

}
\author{
Luke A. Rogers
}
