% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmmFit}
\alias{mmmFit}
\title{Fit a Markovian Movement Model}
\usage{
mmmFit(
  data,
  parameters,
  model,
  random = NULL,
  map = NULL,
  nlminb_control = mmmTMBcontrol()
)
}
\arguments{
\item{data}{A list of named data objects. See details.}

\item{parameters}{A list of initial parameter values. See details.}

\item{model}{A list of values that define the model. See details.}

\item{random}{A character vector of parameters to treat as random effects.}

\item{map}{A list of values to override defaults.}

\item{nlminb_control}{A list. See \code{mmmTMBcontrol()}.}
}
\value{
A list of class \code{mmmFit} and \code{mmmTMB}.
}
\description{
Fit a Markovian Movement Model
}
\details{
The following must be included in the \code{data} list: \itemize{
\item \code{T} An array holding tagged release counts. See
\code{mmmArrayify}. \item \code{R} An array holding recovered counts. See
\code{mmmArrayify}. \item \code{I} A square binary matrix indicating which
direct movement rates to estimate.} The following can be optionally
included in the \code{data} list and are otherwise excluded from the model
\itemize{ \item \code{lambda} An array of tag reporting rates. } The
following can be optionally included in the \code{data} list to be treated
as data, or omitted to be treated as parameters with default initial
values, or included in the \code{parameters} list along with initial values
\itemize{ \item \code{M} An array of instantaneous natural mortality rates.
\item \code{F} An array of instantaneous fishing mortality rates. \item
\code{b} A vector of selectivity values. \item \code{h} A scalar
instantaneous tag loss rate. \item \code{c} A scalar mean initial tag loss
rate.} The following can be included in the \code{model} list, or omitted
to be assigned default values \itemize{ \item \code{liberty} A vector of
length two specifying the minimum and maximum time at liberty. \item
\code{family} An integer specifying the error family 0: Poisson; 1: NB1; 2:
NB2. \item \code{timestep} An integer specifying the results time step as a
multiple of the data time step. \item \code{nlminb_loops} An integer
specifying the number of \code{nlminb} loops. \item \code{newton_steps} An
integer specifying the number of Newton steps. \item \code{openmp_cores} An
integer specifying the number of \code{OpenMP} parallel cores.}
}
\examples{

}
\author{
Luke A. Rogers
}
