% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmmSet}
\alias{mmmSet}
\title{Settings for \code{mmmFit()}}
\usage{
mmmSet(
  error_family = c(nb1 = 1, poisson = 0),
  min_liberty = 1,
  max_liberty = 1000,
  time_varying = 0,
  time_process = c(none = 0, rw = 1),
  cycle_length = 0,
  block_length = 0,
  results_step = 1,
  nlminb_loops = 5,
  newton_iters = 5,
  openmp_cores = 1
)
}
\arguments{
\item{error_family}{\code{\link[=integer]{integer()}} \code{0} = Poisson; \code{1} = NB1;
\code{2} = NB2.}

\item{min_liberty}{\code{\link[=integer]{integer()}} Minimum time steps at liberty before recovery.}

\item{max_liberty}{\code{\link[=integer]{integer()}} Maximum time steps at liberty before recovery.
recapture.}

\item{time_varying}{\code{\link[=integer]{integer()}} \code{0} = None; \code{1} = Time varying
movement rates.}

\item{time_process}{\code{\link[=integer]{integer()}} \code{0} = None; \code{1} = Negative
binomial (NB1).}

\item{cycle_length}{\code{\link[=integer]{integer()}} Cycle length or \code{0} for no cycle.}

\item{block_length}{\code{\link[=integer]{integer()}} Block length for movement rate.}

\item{results_step}{\code{\link[=integer]{integer()}} Results time step as a multiple of the data
time step.}

\item{nlminb_loops}{\code{\link[=integer]{integer()}} Number of times to run \code{\link[stats:nlminb]{stats::nlminb()}}
optimization.}

\item{newton_iters}{\code{\link[=integer]{integer()}} Number of Newton optimization steps.}

\item{openmp_cores}{\code{\link[=integer]{integer()}} Number of parallel cores for TMB.}
}
\value{
A list of settings for \code{mmmFit()}
}
\description{
Settings for \code{mmmFit()}
}
\examples{

}
