% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmmSet}
\alias{mmmSet}
\title{Settings for \code{mmmFit()}}
\usage{
mmmSet(
  error_family = c("nb1", "poisson"),
  time_varying = 0,
  time_process = c("none", "rw"),
  cycle_length = 0,
  block_length = 1,
  fish_rate_by = c("none", "block", "area", "both"),
  results_step = 1,
  nlminb_loops = 5,
  newton_steps = 5,
  openmp_cores = NULL
)
}
\arguments{
\item{error_family}{Character. One of \code{"nb1"} or \code{"poisson"}}

\item{time_varying}{Logical. Should movement rates vary through time?}

\item{time_process}{Character. One of \code{"none"} or \code{"rw"}}

\item{cycle_length}{Integer. Cycle length or \code{0} for no cycle.}

\item{block_length}{Integer. Block length for movement rate.}

\item{fish_rate_by}{Character. Estimate F as a single value (\code{"none"}),
by time steps as blocks (\code{"block"}), by areas (\code{"areas"}), or by
both blocks and areas (\code{"both"}).}

\item{results_step}{Integer. Results time step as a multiple of the data
time step.}

\item{nlminb_loops}{Integer. Number of times to run \code{\link[stats:nlminb]{stats::nlminb()}}
optimization.}

\item{newton_steps}{Integer. Number of Newton optimization steps.}

\item{openmp_cores}{Integer. Number of parallel cores for TMB.}
}
\value{
A list of settings for \code{mmmFit()}
}
\description{
Settings for \code{mmmFit()}
}
\examples{

}
