% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{mmmSim}
\alias{mmmSim}
\title{Simulate Tag Recovery Data}
\usage{
mmmSim(data, parameters = NULL, settings = mmmSet())
}
\arguments{
\item{data}{\code{\link[=list]{list()}} Input data. See details.}

\item{parameters}{\code{\link[=list]{list()}} Parameters to be treated as data.
See details.}

\item{settings}{\code{\link[=list]{list()}} Values that help define the simulation model.
See \code{mmmSet()}.}
}
\value{
An object of class \code{mmmSim}.
}
\description{
Simulate Tag Recovery Data
}
\details{
The \code{\link[=list]{list()}} argument \code{data} must contain:
\itemize{
\item \code{x} An integer matrix of tag release counts formatted
as described in \code{\link[=mmmTags]{mmmTags()}}.
\item \code{y} An integer matrix of tag recovery counts formatted
as described in \code{\link[=mmmTags]{mmmTags()}}.
\item \code{z} A square binary index matrix that indicates allowed
direct movement between areas (from rows to columns) from one
time step to the next. Ones off the diagonal indicate allowed
direct movement. Zeros off the diagonal represent disallowed
direct movement. Numbers on the diagonal are ignored because
self-movement is always allowed. See \code{\link[=mmmIndex]{mmmIndex()}}.
\item \code{h} The (positive) scalar instantaneous tag loss rate.
\item \code{u} The scalar proportion of tags lost during release.
}
The list argument \code{data} may optionally contain:
\itemize{
\item \code{l} A matrix of tag reporting rates (proportions).
Defaults to one if omitted (full reporting). See \code{\link[=mmmRates]{mmmRates()}}.
\item \code{w} A matrix of fishing mortality rate weights. Useful
when the tag time step is finer than the fishing rate step.
Defaults to one (equal fishing rates across tag time steps
within a fishing rate time step). See \code{\link[=mmmWeights]{mmmWeights()}}.
\item \code{f} A matrix of fishing mortality rates. Must be included
in \code{data} or \code{parameters}. See \code{\link[=mmmRates]{mmmRates()}}.
\item \code{m} The scalar instantaneous natural mortality rate.
Must be included in \code{data} or \code{parameters}.
}
The \code{\link[=list]{list()}} argument \code{parameters} must contain:
\itemize{
\item \code{p} An array of movement parameters. See
\code{\link[=create_movement_parameters]{create_movement_parameters()}}.
}
The list argument \code{parameters} may optionally contain:
\itemize{
\item \code{f} A matrix of fishing mortality rates. Must be included
in \code{data} or \code{parameters}. See \code{\link[=mmmRates]{mmmRates()}}.
\item \code{m} The scalar instantaneous natural mortality rate.
Must be included in \code{data} or \code{parameters}.
}
}
\examples{
# Data
sim_x <- create_release_matrix()
sim_z <- mmmIndex(3, pattern = 1)
sim_f <- as.matrix(0.04)
sim_m <- 0.1
sim_h <- 0.02
sim_u <- 0.1
# Parameters
v <- c(0.9, 0.1, 0.0, 0.1, 0.8, 0.3, 0.0, 0.1, 0.7)
sim_r <- array(v, dim = c(3,3,1,1))
sim_p <-create_movement_parameters(sim_r, sim_z)
# Simulation
data <- list(
  x = sim_x,
  z = sim_z,
  f = sim_f,
  m = sim_m,
  h = sim_h,
  u = sim_u)
parameters <- list(p = sim_p)
s1 <- mmmSim(data, parameters)

}
