% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{mmmSim}
\alias{mmmSim}
\title{Simulate Tag Recovery Data}
\usage{
mmmSim(data, parameters, settings = mmmSet())
}
\arguments{
\item{data}{A list.}

\item{parameters}{A list.}

\item{settings}{A list. See \code{mmmSet()}.}
}
\value{
An object of class \code{mmmSim}.
}
\description{
Simulate Tag Recovery Data
}
\examples{

}
