% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{mmmTags}
\alias{mmmTags}
\title{Prepare Tag Data}
\usage{
mmmTags(
  x,
  y,
  cols = NULL,
  groups = NULL,
  step = NULL,
  limits = NULL,
  days_liberty = NULL
)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame()}} Tag release data. See details.}

\item{y}{\code{\link[=data.frame]{data.frame()}} Tag recovery data. See details.}

\item{cols}{\code{\link[=list]{list()}} Named list of character strings. See details.}

\item{groups}{\code{\link[=list]{list()}} Named list of atomic vectors. See details.}

\item{step}{\code{\link[=character]{character()}} One of \code{year}, \code{month}, or
\code{day}.}

\item{limits}{\code{\link[=character]{character()}} Start and end dates as \code{"\%Y-\%M-\%D"}
character strings.}

\item{days_liberty}{\code{\link[=integer]{integer()}} Minimum number of days at liberty.}
}
\value{
A list object of class \code{mmmTags}.
}
\description{
Prepare tag release and recovery data prior to fitting
a Markov movement model via \code{mmmFit()}.
}
\details{
The \code{x} & \code{y} arguments must be data frames that contain
certain columns that have the same column name in both data frames.
Both \code{x} & \code{y} must include:
\itemize{
\item \code{\link[=character]{character()}} column giving unique individual tag IDs.
\item \code{\link[=Date]{Date()}} column giving individual tag release dates.
\item \code{\link[=integer]{integer()}} column giving release areas indexed from one.
}
In addition, \code{y} must include:
\itemize{
\item \code{\link[=Date]{Date()}} column giving individual tag recovery dates.
\item \code{\link[=integer]{integer()}} column giving recovery areas indexed from one.
}
Optionally, both \code{x} & \code{y} can include:
\itemize{
\item \code{\link[=atomic]{atomic()}} column giving values grouped by \code{groups}.
}
Note that \code{\link[=character]{character()}} columns in \code{"\%Y-\%M-\%D"} format may be substituted
for \code{\link[=Date]{Date()}} columns.

The \code{cols} argument is a \code{\link[=list]{list()}} of named elements. The elements
must be \code{\link[=character]{character()}} string names of columns in \code{x} or
\code{y}. There must be one element for each of the required element
names (except as noted):
\itemize{
\item \code{id}: \code{x} & \code{y} Individual tag identification
\item \code{release_date}: \code{x} & \code{y} Release date
\item \code{release_area}: \code{x} & \code{y} Release area
\item \code{recover_date}: \code{y} Recovery date
\item \code{recover_area}: \code{y} Recovery area
\item \code{group}: (Optional) Grouping variable
}

The \code{groups} argument is a \code{\link[=list]{list()}} of named elements. The elements
must be atomic vectors that include all values included in each named
group. All element vectors must be of the same type. The names of the
element vectors must be unique. For \code{groups} to have any effect,
the element \code{group} in argument \code{cols} must name a column
in both \code{x} & \code{y} that has values include in the element
vectors in \code{groups}.
}
\examples{

x <- data.frame(
  Release_Date = c("2010-06-01", "2015-04-01", "2018-06-01", "2019-06-01"),
  Release_Area = c(1,1,1,1),
  Release_Group = c("F", "F", "F", "M"),
  ID = c("TAG001", "TAG002", "TAG003", "TAG004"))
y <- data.frame(
  Release_Date = c("2010-06-01", "2018-06-01", "2019-06-01"),
  Release_Area = c(1,1,1),
  Release_Group = c("F", "F", "M"),
  Recover_Date = c("2013-08-01", "2020-01-01", "2020-08-01"),
  Recover_Area = c(1,2,1),
  ID = c("TAG001", "TAG003", "TAG004"))
cols <- list(
  release_date = "Release_Date",
  release_area = "Release_Area",
  recover_date = "Recover_Date",
  recover_area = "Recover_Area",
  group = "Release_Group",
  id = "ID")
groups <- list(m = "M", f = "F")
step <- "year"
limits <- c("2010-01-01", "2020-12-31")
days_liberty <- 0L

tags <- mmmTags(x, y, cols, groups, step, limits, days_liberty)

}
