% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mmmTMB}
\alias{mmmTMB}
\title{Fit a Markov Movement Model to Mark-Recapture Data}
\usage{
mmmTMB(data_list, ...)
}
\arguments{
\item{released_3d}{\code{\link[=array]{array()}} Release counts by release time, release area,
and class; \code{dim = c(nt, na, ng)}}

\item{recovered_5d}{\code{\link[=array]{array()}} Recovery counts by release time, recovery
time, release area, recovery area, and class;
\code{dim = c(nt, nt, na, na, ng)}}

\item{capture_rate_2d}{\code{\link[=array]{array()}} Instantaneous capture rate per time unit by
time and area; \code{dim = c(nt, na)}}

\item{report_ratio_2d}{\code{\link[=array]{array()}} Reporting ratio by time and area;
\code{dim = c(nt, na)}}

\item{tag_loss_rate}{\code{\link[=numeric]{numeric()}} Instantaneous tag loss rate per time unit,
assumed constant across times, areas, and classes}

\item{imm_loss_ratio}{\code{\link[=numeric]{numeric()}} Ratio of tags lost during release, assumed
constant across times, areas, and classes}

\item{template_2d}{\code{\link[=array]{array()}} Template indicating movement parameters to
estimate; \code{dim = c(na, na)}. See Details.}

\item{recapture_delay}{\code{\link[=integer]{integer()}} Number of time units after release when
recovery becomes possible, \code{>= 1}.}

\item{error_family}{\code{\link[=integer]{integer()}} One of 0: Poisson; 1: NB1; or 2: NB2}

\item{result_units}{\code{\link[=integer]{integer()}} Results time unit as a multiple of the data
time unit. See details.}

\item{time_process}{\code{\link[=integer]{integer()}} Timevarying process for movement
probabilities. One of 0: Constant; 1: RW; 2: AR1.}

\item{time_pattern}{\code{\link[=integer]{integer()}} Timevarying pattern for movement
probabilities. One of 0: Constant; 1: Stepped; 2: Cyclical. See Details.}

\item{pattern_size}{\code{\link[=integer]{integer()}} Number of time units in each
\code{time_pattern} step or cycle.}

\item{newton_steps}{\code{\link[=integer]{integer()}} Number of Newton optimization steps}

\item{nlminb_loops}{\code{\link[=integer]{integer()}} Number of times to run \code{stats::nlminb()}
optimization.}

\item{openmp_cores}{\code{\link[=integer]{integer()}} Number of cores for TMB}

\item{simulate_value}{\code{\link[=logical]{logical()}} Simulate \code{recovered_5d}? See Details.}

\item{capture_map_2d}{\code{\link[=array]{array()}} Custom \code{TMB} map for
\code{log_capture_bias_2d} parameters, \code{dim = c(na, ng)}. See Details.}

\item{structure_list}{\code{\link[=list]{list()}} Input data and model structure constants as
a list. See Details.}

\item{parameter_list}{\code{\link[=list]{list()}} Input (log) initial or simulation parameter
values. See Details.}

\item{optimizer_list}{\code{\link[=list]{list()}} Input optimizer and \code{OpenMP} values. See
Details.}

\item{nlminb_control}{\code{\link[=list]{list()}} See \code{mmmTMBcontrol()}}
}
\value{
An \code{list()} of class \code{mmmTMB} with elements:
\itemize{
\item{\code{adfun}: The TMB AD model object}
\item{\code{model}: Output from \code{nlminb()}}
\item{\code{structure}: A \code{list()} of data and constants}
\item{\code{parameter}: A \code{list()} of initial (log) parameter values}
\item{\code{optimizer}: A \code{list()} of optimizer constants}
\item{\code{results}: A \code{list()} of results data frames}
\item{\code{map}: A \code{list()} of parameter map \code{factor()}s}
\item{\code{map}: A \code{list()} of random effects}
\item{\code{sd_report}: A \code{list()} generated by \code{sdreport()}}
\item{\code{conv}: A \code{list()} of convergence diagnostics}
\item{\code{mgc}: The maximum absolute gradient component}
}
}
\description{
Estimate movement probabilities among areas by fitting a Markov
movement model to mark release and recovery data.

Array dimensions are abbreviated for space and are given by:
\itemize{
\item{\code{nt}: Count of time units in the data}
\item{\code{na}: Count of areas in the data}
\item{\code{ng}: Count of class categories}
}
}
\details{
Describe \code{template_2d}...
Describe \code{result_units}...
Describe \code{time_pattern}...
Describe \code{capture_map_2d}...
Describe \code{structure_list}...
Describe \code{parameter_list}...
Describe \code{optimizer_list}...
Describe \code{simulate_value}...
}
\examples{

}
